+++ src/main/scala/advent2019/Day7.scala
@@ -51,7 +51,7 @@ object Day7 extends Day(7){
         for (i <- 0 to 4) {
           val result = OpcodeProcessor.processDay5OppCode(
             amplifiers.getOrElse(i, Program(program.clone(), 0)),
-            if(amplifiers.contains(i)) List[Int](output) else List[Int](ampVals(i), output))
+            if(amplifiers.contains(i)) List[Long](output) else List[Long](ampVals(i), output))
           amplifiers = amplifiers + (i -> result)
           output = result.output.reverse.head.toInt
         }
